Created new emailing class for the same reason
