+++ src/main/java/codeu/controller/ChatServlet.java
@@ -24,16 +24,10 @@ import codeu.model.store.basic.MessageStore;
 import codeu.model.store.basic.UserStore;
 import codeu.utils.ImageStorage;
 import codeu.utils.TextFormatter;
+import codeu.utils.Email;
 import java.io.IOException;
-import java.io.UnsupportedEncodingException;
 import java.time.Instant;
 import java.util.*;
-import javax.mail.MessagingException;
-import javax.mail.Session;
-import javax.mail.Transport;
-import javax.mail.internet.AddressException;
-import javax.mail.internet.InternetAddress;
-import javax.mail.internet.MimeMessage;
 import javax.servlet.ServletException;
 import javax.servlet.http.*;
 import org.jsoup.Jsoup;
@@ -174,7 +168,7 @@ public class ChatServlet extends HttpServlet {
     } else if (button == null && conversation.getConversationUsers().contains(user.getId())) {
       createMessage(request, user, conversation);
 
-    } 
+    }
     // redirect to a GET request
     response.sendRedirect("/chat/" + conversationTitle);
   }
@@ -296,7 +290,8 @@ public class ChatServlet extends HttpServlet {
                     conversation.getIsPublic());
     activityStore.addActivity(activity);
 
-    sendEmailNotification(user, conversation);
+    Email email = new Email();
+    email.sendEmailNotification(user, conversation);
 
   }
 
@@ -338,62 +333,5 @@ public class ChatServlet extends HttpServlet {
                     conversation.getConversationUsers(),
                     conversation.getIsPublic());
     activityStore.addActivity(activity);
-  }
-
-  /**
-   * Method to send an email notification to all users in a conversation who are not logged on other
-   * than the message sender
-   */
-  public void sendEmailNotification(User user, Conversation conversation) {
-
-    Properties props = new Properties();
-    Session session = Session.getDefaultInstance(props, null);
-
-    List<UUID> conversationUsers = conversation.getConversationUsers();
-
-    String msgBody =
-            user.getName()
-                    + " sent a message in "
-                    + conversation.getTitle()
-                    + " on "
-                    + conversation.getCreationTime()
-                    + " while you were away. \n \n ";
-
-    SessionListener currentSession = SessionListener.getInstance();
-
-    for (UUID conversationUserUUID : conversationUsers) {
-      User conversationUser = userStore.getUser(conversationUserUUID);
-      if (conversationUser != user
-              && conversationUser != null
-              && !currentSession.isLoggedIn(conversationUser.getName())
-              && conversationUser.getNotificationStatus()) {
-        if (user.getNotificationFrequency().equals("everyMessage")) {
-          try {
-            javax.mail.Message msg = new MimeMessage(session);
-            msg.setFrom(
-                    new InternetAddress(
-                            "chatMessageAdmin@chatu-196017.appspotmail.com", "CodeU Team 12 Admin"));
-            msg.addRecipient(
-                    javax.mail.Message.RecipientType.TO,
-                    new InternetAddress(conversationUser.getEmail(), conversationUser.getName()));
-            msg.setSubject(user.getName() + " has sent you a message");
-            msgBody += " Please log in to view this message";
-            msg.setText(msgBody);
-            Transport.send(msg);
-          } catch (AddressException e) {
-            System.out.println("Invalid email address formatting. Email not sent.");
-            System.out.println("AddressException:" + e);
-          } catch (MessagingException e) {
-            System.out.println("An error has occurred with this message. Email not sent.");
-            System.out.println("MessagingException:" + e);
-          } catch (UnsupportedEncodingException e) {
-            System.out.println("This character encoding is not supported. Email not sent");
-            System.out.println("UnsupportedEncodingException:" + e);
-          }
-        }
-      } else {
-        user.addNotification(msgBody);
-      }
-    }
   }
 }
